# api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PudimFinance API",
    "description": "Personal finance application API. Immutable double-entry ledger, event sourcing, real-time reporting, receipt scanning and price tracking.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/summary": {
      "get": {
        "tags": [
          "Summary"
        ],
        "summary": "Returns monthly totals with a category breakdown.",
        "operationId": "get_summary",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "description": "Year to summarise (default: current)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "query",
            "description": "Month 1-12 (default: current)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Monthly summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid month parameter"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategorySummary": {
        "type": "object",
        "description": "Per-category totals for the summary endpoint.",
        "required": [
          "total"
        ],
        "properties": {
          "category_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Category UUID (null for transactions without a category)."
          },
          "category_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Category display name."
          },
          "color": {
            "type": [
              "string",
              "null"
            ],
            "description": "Hex color code for the category."
          },
          "icon": {
            "type": [
              "string",
              "null"
            ],
            "description": "Icon identifier for the category."
          },
          "total": {
            "type": "string",
            "description": "Total amount for this category."
          }
        }
      },
...
      "SummaryResponse": {
        "type": "object",
        "description": "Monthly summary response with income, expense, balance, and category breakdown.",
        "required": [
          "income_total",
          "expense_total",
          "balance",
          "by_category",
          "year",
          "month"
        ],
        "properties": {
          "balance": {
            "type": "string",
            "description": "Balance = income − expense."
          },
          "by_category": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategorySummary"
            },
            "description": "Per-category *expense* breakdown for the month (income accounts are\nreported through `income_total`)."
          },
          "expense_total": {
            "type": "string",
            "description": "Total expenses for the selected month (positive value)."
          },
          "income_total": {
            "type": "string",
            "description": "Total income for the selected month (positive value)."
          },
          "month": {
            "type": "integer",
            "format": "int32",
            "description": "Month used for the query (1-12).",
            "minimum": 0
          },
          "year": {
            "type": "integer",
            "format": "int32",
            "description": "Year used for the query."
          }
        }
      },
...
    }
  },
  "security": [
    {}
  ],
...
}